chore: update plugin docusaurus ts types (#432)
* chore: update plugin-docusaurus types

* chore: file type

* chore: version and changelog
