+++ packages/plugin-docusaurus/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@ice/pkg-plugin-docusaurus",
-  "version": "1.4.1",
+  "version": "1.4.2",
   "description": "Plugin for component previewing",
   "main": "es2017/index.mjs",
   "exports": {
@@ -34,8 +34,8 @@
     "prepublishOnly": "pnpm run-s lint build"
   },
   "dependencies": {
-    "@docusaurus/core": "^2.0.0",
-    "@docusaurus/preset-classic": "^2.0.0",
+    "@docusaurus/core": "^2.2.0",
+    "@docusaurus/preset-classic": "^2.2.0",
     "@mdx-js/react": "^1.6.22",
     "@swc/helpers": "^0.4.3",
     "address": "^1.2.1",
